requirements.txt
streamlit
langchain
langchain_groq
langchain_community
python-dotenv
numexpr
wikipedia